int -> size_t in common.hpp

+++ tasks/levonychev_i_min_val_rows_matrix/common/include/common.hpp
@@ -8,9 +8,9 @@
 
 namespace levonychev_i_min_val_rows_matrix {
 
-using InType = std::tuple<std::vector<double>, int, int>;
+using InType = std::tuple<std::vector<double>, size_t, size_t>;
 using OutType = std::vector<double>;
-using TestType = std::tuple<int, int>;
+using TestType = std::tuple<size_t, size_t>;
 using BaseTask = ppc::task::Task<InType, OutType>;
 
 }  // namespace levonychev_i_min_val_rows_matrix
